+++ app/utils.py
@@ -6,10 +6,11 @@ from typing import TYPE_CHECKING
 import aioredis
 from tortoise.exceptions import DoesNotExist
 
+from app.pipedrive.api import get_and_create_or_update_pd_deal
 from app.settings import Settings
 
 if TYPE_CHECKING:
-    from app.models import Config
+    from app.models import Config, CustomField, CustomFieldValue
 
 settings = Settings()
 logger = logging.getLogger('utils')
@@ -69,3 +70,38 @@ def sanitise_string(input_string: str) -> str:
 
     # Convert to lowercase
     return sanitized_string.lower()
+
+
+async def update_or_create_inherited_deal_custom_field_values(company):
+    """
+    Update the inherited custom field values of all company deals with the custom field values of the company
+    then send to pipedrive to update the deal
+    """
+    deal_custom_fields = await CustomField.filter(linked_object_type='Deal')
+    deal_custom_field_machine_names = [cf.machine_name for cf in deal_custom_fields]
+    company_custom_fields_to_inherit = (
+        await CustomField.filter(linked_object_type='Company', machine_name__in=deal_custom_field_machine_names)
+        .exclude(machine_name='hermes_id')
+        .prefetch_related('values')
+    )
+
+    deals = await company.deals
+
+    for cf in company_custom_fields_to_inherit:
+        deal_cf = next((dcf for dcf in deal_custom_fields if dcf.machine_name == cf.machine_name), None)
+        if not deal_cf:
+            continue
+
+        if cf.values:
+            value = cf.values[0].value
+        elif cf.hermes_field_name:
+            value = getattr(company, cf.hermes_field_name, None).id
+        else:
+            raise ValueError(f'No value for custom field {cf}')
+
+        for deal in deals:
+            await CustomFieldValue.update_or_create(
+                **{'custom_field_id': deal_cf.id, 'deal': deal, 'defaults': {'value': value}}
+            )
+
+            await get_and_create_or_update_pd_deal(deal)
