+++ tests/test_combined.py
@@ -639,7 +639,6 @@ class TestDealCustomFieldInheritance(HermesTestCase):
         create org
         create deal
         """
-        debug('----------------------------test_com_cli_create_update_org_deal---------------------------------------')
         mock_gcal_builder.side_effect = fake_gcal_builder()
         mock_pd_request.side_effect = fake_pd_request(self.pipedrive)
 
@@ -670,7 +669,6 @@ class TestDealCustomFieldInheritance(HermesTestCase):
             sells_payg=True,
             sells_gb=True,
         )
-        debug(sales_person.id)
         assert await Company.all().count() == 0
         assert await Contact.all().count() == 0
         r = await self.client.post(self.callbooker_callback, json={'admin_id': sales_person.id, **CB_MEETING_DATA})
@@ -708,7 +706,6 @@ class TestDealCustomFieldInheritance(HermesTestCase):
         assert await deal.company == company
         assert await deal.admin == sales_person
 
-        debug(deal.__dict__)
         await pd_post_process_sales_call(company, contact, meeting, deal)
 
         assert self.pipedrive.db['organizations'] == {
@@ -721,8 +718,6 @@ class TestDealCustomFieldInheritance(HermesTestCase):
                 '123_sales_person_456': sales_person.id,
             }
         }
-        debug(sales_person.id)
-        debug(sales_person.pd_owner_id)
         assert self.pipedrive.db['deals'] == {
             1: {
                 'title': 'Junes Ltd',
